move domain limit to DomainSerializer.validate() and update error codes
This commit moves the domain limit validation to the serializer so that
domain creations that do not use a view but do use this serializer will
be able to take this check into account.

The distinction of 409 for "name-unavailable" and 403 for "domain-limit"
does not make sense because both validation errors can occur at the same
time. This commit thus removes the special treatment and falls back to
400 for these cases.